+++ src/Log4JParserCpp/Iterator.h
@@ -20,7 +20,7 @@ namespace Log4JParser
     public:
         bool MoveNext ();
 
-        Event Current (int32_t *id = nullptr) const;
+        Event Current (size_t *id = nullptr) const;
 
     protected:
         static const Log4JIterator *GetIterator (const IteratorBase *iterator);
